Start on preprocessing script
preprocess.py takes the path to annotation.txt and an output path and writes cropped and resized images to the output path.

This requires python 3 (which has types now?!) and tf-nightly (because of a bug https://stackoverflow.com/a/47720915)

This is just a start. Ideally this will serialize these out as numpy arrays or some tf binary format rather than jpgs.